admin_panel/requirements.txt
arrow==1.3.0
asgiref==3.7.2
bleach==6.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
dj-mongo-backend==0.1.4
Django==4.2.7
django-debug-toolbar==4.2.0
django-summernote==0.8.20.0
dnspython==2.4.2
docker==6.1.3
environs==9.5.0
idna==3.6
Jinja2==3.1.2
jinja2-time==0.2.0
make==0.1.6.post2
marshmallow==3.20.1
MarkupSafe==2.1.3
marshmallow==3.20.1
mongoengine==0.27.0
packaging==23.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.8.0
webencodings==0.5.1
typer==0.9.0
typing_extensions==4.8.0
requests==2.31.0
pyparsing==3.0.9
tzdata==2022.1
psycopg2-binary==2.9.9
gunicorn==20.0.4
types-python-dateutil==2.8.19.14
urllib3==2.1.0
webencodings==0.5.1
websocket-client==1.7.0
